requirements.txt
Django==1.6.5
South==1.0
httplib2==0.9
psycopg2==2.5.3
python-instagram==1.1.3
requests==2.3.0
simplejson==3.6.2
stripe==1.19.0
wsgiref==0.1.2
